using read/write lock to sync between threads
I've tried to change the conccurrent dictionary to a simple dictionarry and add a ReaderWriterLockSlim to sync between threads. Results: no improvements.

The first tought was to use a dictionary for each thread, then combine them on read, but I couldn't find a simple and fast way to ensure that players will be updated by the same thread every time.