Tested with an empty PlayerManager
With this test, I tried to completely remove the concurrency aspect of the app. The conclusion is that concurrency was not the problem, but rather the amount of data that had to be transmitted.

To update 1000 players with information about all 1000 players, I would have to transmit at least 24MB of data per update frame:
- 8 bytes: double for X position
- 8 bytes: double for Y position
- at least 4 bytes: string for Name (if it's empty, which certainly won't be).
- and all of that 1000 * 1000 times (connections * num of updates) per 'frame'.